remove jitter and add os_random in libnetdata
