+++ CMakeLists.txt
@@ -862,8 +862,8 @@ set(LIBNETDATA_FILES
         src/libnetdata/log/nd_log-to-windows-common.h
         src/libnetdata/common.h
         src/libnetdata/xxHash/xxhash.h
-        src/libnetdata/os/jitter.c
-        src/libnetdata/os/jitter.h
+        src/libnetdata/os/random.c
+        src/libnetdata/os/random.h
 )
 
 if(ENABLE_PLUGIN_EBPF)
