+++ src/libnetdata/clocks/clocks.h
@@ -4,7 +4,7 @@
 #define NETDATA_CLOCKS_H 1
 
 #include "../libnetdata.h"
-#include "libnetdata/os/jitter.h"
+#include "libnetdata/os/random.h"
 
 #ifndef HAVE_CLOCK_GETTIME
 struct timespec {
@@ -32,7 +32,6 @@ typedef struct heartbeat {
     usec_t randomness;
     size_t statistics_id;
     XXH64_hash_t hash;
-    OS_JITTER jitter;
 } heartbeat_t;
 
 /* Linux value is as good as any other */
