+++ src/streaming/sender.c
@@ -506,8 +506,6 @@ void *rrdpush_sender_thread(void *ptr) {
     bool was_connected = false;
     size_t iterations = 0;
     time_t now_s = now_monotonic_sec();
-    OS_JITTER jitter;
-    os_jitter_init(&jitter, os_gettid());
 
     while(!rrdhost_sender_should_exit(s)) {
         iterations++;
@@ -619,9 +617,6 @@ void *rrdpush_sender_thread(void *ptr) {
                 .revents = 0,
             }
         };
-
-        // let it queue for up to 1ms
-        os_jitter_wait(&jitter, USEC_PER_MS);
 
         int poll_rc = poll(fds, 2, 50); // timeout in milliseconds
